rowexec: fix unmatched row bug for second paired join
If a group is umatched in the second join in a paired join,
and a row needs to be output for the group, we were not
setting the right-side columns that are in the left-side to
NULL, since the expectation was that the NULL right-side
columns would be the ones kept after the subsequent projection.
This assumption was incorrect: the projection keeps the
right-side key columns that are included in the left-side.
The fix is to NULL these columns.

Release note: None