[DNM] storage,server: enable separated intents
Separated intents are enabled by default, and the setting
is randomized on the testing path for testserver.go.

The intent resolution performance, which caused these
to be disabled, has been improved due to
https://github.com/cockroachdb/cockroach/pull/60622 and
https://github.com/cockroachdb/cockroach/pull/60698

Informs #41720

Release note (ops change): The default value of the
storage.transaction.separated_intents.enabled cluster
setting is changed to true.